homework2/requirements.txt
asgiref==3.7.2
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.11
django-bootstrap5==24.3
djangorestframework==3.15.2
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
PySocks==1.7.1
pytz==2024.1
requests==2.32.3
selenium==4.29.0
setuptools==68.1.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
supervisor==4.2.5
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wheel==0.42.0
wsproto==1.2.0
